Yet more cleanup
